SimilarityMap/requirements.txt
psycopg2-binary
numpy
pandas
umap-learn
python-dotenv